pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "swing-trader"
version = "0.1.0"
description = "日本株スイングトレード自動スクリーニングシステム"
requires-python = ">=3.11"
dependencies = [
    "jquants-api-client>=2.0",
    "pandas>=2.0",
    "pandas_ta>=0.3.14",
    "numpy>=1.24",
    "requests>=2.31",
    "yfinance>=0.2.31",
    "beautifulsoup4>=4.12",
    "lxml>=5.0",
    "feedparser>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
